xfsdump: rename worker threads
xfsdump's use of "slave," while largely understood in technical circles,
poses a barrier for inclusion to some potential members of the development
and user community, due to the historical context of masters and slaves,
particularly in the United States.

As these entities are already controlled by "managers" in the code,
simply rename them to "workers," which is a more natural fit in any
case.

Signed-off-by: Eric Sandeen <sandeen@redhat.com>
Reviewed-by: Darrick J. Wong <darrick.wong@oracle.com>
Reviewed-by: Pavel Reichl <preichl@redhat.com>
Signed-off-by: Eric Sandeen <sandeen@sandeen.net>